+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 0.13.0
+- [BRAVE] Add brave support
+- [CHROME] Fix expires handling
+- [CHROME] Fix double quoted sqlite strings
+### 0.11.4
+- [CHROME] Support Chromium v80
+- [CHROME] Better error when missing cookie file
 ### 0.11.3
 - [FIREFOX] Add only cookies from the specified domain
 ### 0.11.2